Handle apps being run on other ports for dev
Before this commit the configuration didn't handle setting the APP_URL
to run on a specific port. This also changes the name of some variables
in the .env and updates and clean up some files in the build folder.
Updates readme to be more explicit about the .env file and its
variables.